Change slot.assign() from sequence to variadic
This spec change wasn't noticed when [1] was landed. This CL aligns
the implementation and WPTs with the spec IDL file [2]. As with [1],
the use counter [3] is *very* low, so there should not be compat
concerns with this change.

I have requested a re-review of the I2S, at [4]. I will wait to ship
this until I receive 3 more LGTMs there.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2824763
[2] https://html.spec.whatwg.org/#htmlslotelement
[3] https://chromestatus.com/metrics/feature/timeline/popularity/3442
[4] https://groups.google.com/a/chromium.org/g/blink-dev/c/6U78F3KWJ78/m/CExRJ9KlBQAJ

Fixed: 1202591
Bug: 1196842
Change-Id: If0e21d2f3329b9903a9ffeb0df4a80143f6e1581